GH-2695 Nest scope in ExplosionControlConfiguration, drop javadocs
Move ExplosionControlScope to a nested `ExplosionControlConfiguration.Scope`
(mirrors the TntProtectionConfiguration nested-config pattern) and remove the
javadoc blocks; the @Comment config docs are kept.
