+++ plugin/src/main/java/net/dzikoysk/funnyguilds/config/ExplosionControlConfiguration.java
@@ -2,39 +2,105 @@ package net.dzikoysk.funnyguilds.config;
 
 import eu.okaeri.configs.OkaeriConfig;
 import eu.okaeri.configs.annotation.Comment;
+import eu.okaeri.configs.annotation.CustomKey;
+import eu.okaeri.configs.annotation.Exclude;
 import eu.okaeri.configs.annotation.NameModifier;
 import eu.okaeri.configs.annotation.NameStrategy;
 import eu.okaeri.configs.annotation.Names;
+import eu.okaeri.validator.annotation.Min;
+import java.util.Collections;
+import java.util.EnumMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import org.bukkit.Material;
+import org.jetbrains.annotations.Nullable;
 
-/**
- * Controls how explosions destroy blocks, split into a {@code guild} scope (explosions on guild territory)
- * and a {@code global} scope (explosions elsewhere). The active scope is chosen by the explosion's location.
- */
 @Names(strategy = NameStrategy.HYPHEN_CASE, modifier = NameModifier.TO_LOWER_CASE)
 public class ExplosionControlConfiguration extends OkaeriConfig {
 
     @Comment("Wybuchy na terenie gildii")
-    public ExplosionControlScope guild = defaultScope();
+    public Scope guild = defaultScope();
 
     @Comment("")
     @Comment("Wybuchy poza terenem gildii")
-    public ExplosionControlScope global = defaultScope();
+    public Scope global = defaultScope();
 
     public void loadProcessedProperties() {
         this.guild.loadProcessedProperties();
         this.global.loadProcessedProperties();
     }
 
-    private static ExplosionControlScope defaultScope() {
+    private static Scope defaultScope() {
         Map<String, Double> materials = new LinkedHashMap<>();
         materials.put("ender_chest", 20.0);
         materials.put("enchantment_table", 20.0);
         materials.put("obsidian", 20.0);
         materials.put("water", 33.0);
         materials.put("lava", 33.0);
-        return ExplosionControlScope.of(3, -1.0, materials);
+        return Scope.of(3, -1.0, materials);
+    }
+
+    @Names(strategy = NameStrategy.HYPHEN_CASE, modifier = NameModifier.TO_LOWER_CASE)
+    public static class Scope extends OkaeriConfig {
+
+        @Min(0)
+        @Comment("Zasięg pobieranych przedmiotów po wybuchu, jeżeli chcesz wyłączyć - wpisz 0")
+        public int radius = 3;
+
+        @Comment("")
+        @Comment("Zachowanie dla materiałów spoza listy 'materials':")
+        @Comment("-1  = zachowanie domyślne (vanilla)")
+        @Comment(" 0  = nie niszcz nic poza listą oraz pomiń bloki niszczone domyślnie przez wybuch (ochrona terenu)")
+        @Comment(">0  = szansa (w %) na zniszczenie każdego innego materiału")
+        @CustomKey("default")
+        public double defaultChance = -1.0;
+
+        @Comment("")
+        @Comment("Materiały i ich indywidualna szansa zniszczenia (w %)")
+        public Map<String, Double> materials = new LinkedHashMap<>();
+
+        @Exclude
+        private Map<Material, Double> resolvedMaterials = Collections.emptyMap();
+
+        public void loadProcessedProperties() {
+            Map<Material, Double> resolved = new EnumMap<>(Material.class);
+            this.materials.forEach((name, chance) -> {
+                if (chance == null || chance < 0) {
+                    return;
+                }
+
+                Material material = Material.matchMaterial(name);
+                if (material != null && material != Material.AIR) {
+                    resolved.put(material, chance);
+                }
+            });
+            this.resolvedMaterials = resolved;
+        }
+
+        public int getRadius() {
+            return this.radius;
+        }
+
+        public boolean dropsVanillaBlocks() {
+            return this.defaultChance == 0.0;
+        }
+
+        public @Nullable Double explosionChance(Material material) {
+            Double chance = this.resolvedMaterials.get(material);
+            if (chance != null) {
+                return chance;
+            }
+            return this.defaultChance > 0.0 ? this.defaultChance : null;
+        }
+
+        static Scope of(int radius, double defaultChance, Map<String, Double> materials) {
+            Scope scope = new Scope();
+            scope.radius = radius;
+            scope.defaultChance = defaultChance;
+            scope.materials = materials;
+            return scope;
+        }
+
     }
 
 }
